get_growthrate method
